requirements.txt
# Core data science packages
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
scipy==1.10.1
statsmodels==0.14.0

# Time series forecasting
prophet==1.1.4
pmdarima==2.0.3

# NLP and text processing
nltk==3.8.1
spacy==3.6.1
transformers==4.30.2

# Visualization
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.15.0
dash==2.10.2
streamlit==1.23.1

# Data engineering
dbt-core==1.5.0
dbt-postgres==1.5.0
apache-airflow==2.6.3
great-expectations==0.16.13

# AWS packages
boto3==1.26.151
awswrangler==3.2.0

# Database connectors
psycopg2-binary==2.9.6
SQLAlchemy==2.0.16

# Jupyter and development
jupyter==1.0.0
pytest==7.3.1
black==23.3.0
flake8==6.0.0

# Documentation
sphinx==7.0.1
nbsphinx==0.9.2 